🌸 Gardening: cleaning the types part
