+++ completion/typesprovider/openai/types.openai.go
@@ -1,13 +1,16 @@
-package llm
+package openai
 
 // OpenAI API support
-import "encoding/json"
+import (
+	"encoding/json"
+	"github.com/parakeet-nest/parakeet/llm"
+)
 
 // https://platform.openai.com/docs/api-reference/chat/create
 
-type OpenAIQuery struct {
-	Model    string    `json:"model"`
-	Messages []Message `json:"messages"`
+type Query struct {
+	Model    string        `json:"model"`
+	Messages []llm.Message `json:"messages"`
 
 	//Options  OpenAIOptions `json:"options"`
 	//--------------------------------------------
@@ -32,7 +35,7 @@ type OpenAIQuery struct {
 
 	Stream bool `json:"stream"`
 
-	Tools      []Tool `json:"tools,omitempty"`
+	Tools      []llm.Tool `json:"tools,omitempty"`
 	ToolChoice string `json:"tool_choice,omitempty"`
 
 	ParallelToolCalls bool   `json:"parallel_tool_calls,omitempty"` // not used right now
@@ -50,7 +53,7 @@ type OpenAIQuery struct {
 	OpenAIAPIKey string `json:"-"`
 }
 
-func (query *OpenAIQuery) ToJsonString() string {
+func (query *Query) ToJsonString() string {
 	// for the verbose mode
 	// Marshal the answer into JSON
 	jsonBytes, err := json.MarshalIndent(query, "", "  ")
@@ -66,11 +69,9 @@ func (query *OpenAIQuery) ToJsonString() string {
 type OpenAIMessage struct {
 	Role    string `json:"role,omitempty"`
 	Content string `json:"content,omitempty"`
-	//ToolCalls []interface{} `json:"tool_calls,omitempty"` 
-	ToolCalls []map[string]interface{} `json:"tool_calls,omitempty"` 
-
+	//ToolCalls []interface{} `json:"tool_calls,omitempty"`
+	ToolCalls []map[string]interface{} `json:"tool_calls,omitempty"`
 }
-
 
 /*
 type OpenAIToolCall struct {
@@ -84,16 +85,16 @@ type OpenAIToolCall struct {
 //! ????
 
 /*
-        "tool_calls": [
-          {
-            "id": "call_hkPjBb3TnBg532I7LStxDuqr",
-            "type": "function",
-            "function": {
-              "name": "hello",
-              "arguments": "{\"name\":\"Bob\"}"
-            }
-          }
-        ],
+   "tool_calls": [
+     {
+       "id": "call_hkPjBb3TnBg532I7LStxDuqr",
+       "type": "function",
+       "function": {
+         "name": "hello",
+         "arguments": "{\"name\":\"Bob\"}"
+       }
+     }
+   ],
 */
 
 type Delta struct {
@@ -126,7 +127,7 @@ type Usage struct {
 {"id":"chatcmpl-123","object":"chat.completion.chunk","created":1694268190,"model":"gpt-4o-mini", "system_fingerprint": "fp_44709d6fcb", "choices":[{"index":0,"delta":{},"logprobs":null,"finish_reason":"stop"}]}
 */
 
-type OpenAIAnswer struct {
+type Answer struct {
 	ID                string   `json:"id"`
 	Object            string   `json:"object"`
 	Created           int64    `json:"created"`
@@ -136,7 +137,7 @@ type OpenAIAnswer struct {
 	Usage             Usage    `json:"usage"`
 }
 
-func (answer *OpenAIAnswer) ToJsonString() string {
+func (answer *Answer) ToJsonString() string {
 	// for the verbose mode
 	// Marshal the answer into JSON
 	jsonBytes, err := json.MarshalIndent(answer, "", "  ")
@@ -147,59 +148,4 @@ func (answer *OpenAIAnswer) ToJsonString() string {
 	// Convert JSON bytes to string
 	jsonString := string(jsonBytes)
 	return jsonString
-}
-
-/*
-curl https://api.openai.com/v1/embeddings \
-  -H "Content-Type: application/json" \
-  -H "Authorization: Bearer $OPENAI_API_KEY" \
-  -d '{
-    "input": "Your text string goes here",
-    "model": "text-embedding-3-small"
-  }'
-*/
-
-// https://platform.openai.com/docs/guides/embeddings/what-are-embeddings
-type OpenAIQuery4Embedding struct {
-	Input string `json:"input"`
-	Model string `json:"model"`
-
-	OpenAIAPIKey string `json:"-"`
-}
-
-/*
-{
-  "object": "list",
-  "data": [
-    {
-      "object": "embedding",
-      "index": 0,
-      "embedding": [
-        -0.006929283495992422,
-        -0.005336422007530928,
-        ... (omitted for spacing)
-        -4.547132266452536e-05,
-        -0.024047505110502243
-      ],
-    }
-  ],
-  "model": "text-embedding-3-small",
-  "usage": {
-    "prompt_tokens": 5,
-    "total_tokens": 5
-  }
-}
-*/
-
-type OpenAIEmbeddingResponse struct {
-	Object string      `json:"object"`
-	Data   []Embedding `json:"data"`
-	Model  string      `json:"model"`
-	Usage  Usage       `json:"usage"`
-}
-
-type Embedding struct {
-	Object    string    `json:"object"`
-	Index     int       `json:"index"`
-	Embedding []float64 `json:"embedding"`
 }
