+++ embeddings/typesprovider/README.md
@@ -1,0 +1,2 @@
+# Types provider
+> This is a work in progress and subject to change
